requirements.txt
# Core dependencies for Assignment 4: A2A Multi-Agent System

# Google Agent Development Kit
google-adk>=1.9.0

# A2A SDK
a2a-sdk>=0.3.0

# HTTP client for A2A communication
httpx>=0.27.0

# ASGI server for agent servers
uvicorn>=0.30.0

# Environment variable management
python-dotenv>=1.0.0

# Async event loop compatibility
nest-asyncio>=1.6.0

# MCP SDK (also in mcp_server/requirements.txt)
mcp>=1.6.0
